Добавление функционала для загрузки файлов: создание класса FileStorage, контроллера FilesController и репозитория FileRepository; обновление методов для обработки загрузки и удаления файлов с валидацией данных.

+++ Services/AdminService.php
@@ -110,8 +110,14 @@ class AdminService
         return ['id' => $user->id, 'email' => $user->email, 'role' => $user->role, 'age' => $user->age, 'gender' => $user->gender];
     }
 
-    public function deleteUserById(string $id): void
+    public function deleteUserById(string $id): Response
     {
+        if ((int) $id === Config::getConfig('app.idUserSystem')) {
+            return new JSONResponse(Helper::showError('Нельзя удалять системного пользователя'));
+        }
+
         App::getService('adminRepository')::deleteUserBy(['id' => $id]);
+
+        return new JSONResponse();
     }
 }
